活动时长记录与用户中心查询

+++ volunteer-platform/src/main/java/com/student/webproject/user/dto/UserDetailDTO.java
@@ -1,10 +1,11 @@
 package com.student.webproject.user.dto;
 
 import lombok.Data;
+import java.math.BigDecimal; // 1. 导入 BigDecimal
 
 @Data
 public class UserDetailDTO {
-    private Integer id;
+    private Long id; // 2. 类型改为 Long
     private String username;
     private String realName;
     private String email;
@@ -12,5 +13,7 @@ public class UserDetailDTO {
     private String studentId;
     private String avatarUrl;
     private String role;
-    private Double totalServiceHours;
+
+    // 3. 【核心修复】补上总志愿时长字段，类型为 BigDecimal
+    private BigDecimal totalServiceHours;
 }
